modified to use intptr_t instead of long to fix error when compiling on 64-bit platforms (long is 32-bit on windows64)

+++ interfaces/CppSound.cpp
@@ -180,9 +180,9 @@ void CppSound::write(const char *text)
   Message("%s", text);
 }
 
-long CppSound::getThis()
+intptr_t CppSound::getThis()
 {
-  return (long) this;
+  return (intptr_t) this;
 }
 
 CsoundFile *CppSound::getCsoundFile()
